Add test to verify that upgrading the wrapper doesn't break
This can happen on Windows because cmd.exe processes batch files
line-by-line, and editing it in the middle of that can result in
problems. Unfortunately, if this happens, we can't do much about it but
try to make the lines match up again.
