+++ platforms/core-runtime/wrapper-main/src/crossVersionTest/groovy/org/gradle/integtests/wrapper/Wrapper8PlusCrossVersionIntegrationTest.groovy
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2026 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.gradle.integtests.wrapper
+
+import org.gradle.integtests.fixtures.TargetVersions
+
+@TargetVersions("8.0.0+")
+@SuppressWarnings("IntegrationTestFixtures")
+class Wrapper8PlusCrossVersionIntegrationTest extends AbstractWrapperCrossVersionIntegrationTest {
+    void canUseWrapperFromPreviousVersionToUpgradeToCurrentVersionWrapper() {
+        when:
+        def executer = prepareWrapperExecuter(previous, current)
+
+        then:
+        executer.withTasks('wrapper').run()
+        checkWrapperWorksWith(executer, current)
+
+        cleanup:
+        cleanupDaemons(executer, current)
+    }
+}
